Initial code import
This is a partially functioning prototype.

There is an interface, VolumeDriver, that abstracts the Docker
interface for volume management.

The service properly listens on /var/run/blocker/blocker.sock, and
responds to the relevant Docker plugin HTTP REST calls.

A single implementation of VolumeDriver, EbsVolumeDriver, exists.
It does reasonable things in response to the incoming Docker calls,
including maintaining a map from "name" (in the case of EBS, the
volume ID) to "mountpoint" (which in our case isn't really implemented
just yet).  It'll validate arguments and so on.  The interaction with
AWS just started, and it will at least issue errors if you try to
run this from outside an AWS instance and/or specify an invalid volume
ID.  From here, we need to attach, detach, mount, and so on ...