Change the way we find a free device
Any checks we do to find a free device -- either consulting AWS or
looking at the local devices -- is subject to TOCTOU.  Although
that might be a good idea to speed up certain checks, and avoid
needless networking roundtrips, we need to tolerate AWS reporting
that a device is already in use.  So this change simply fires away
and tries each device until AWS says it is okay.