A few attach and mount fixes
On modern Linux kernels, /dev/sd* is known as /dev/xvd*.  We now
check for this and remap the device name as appropriate.

Use the mount and umount commands, as the syscalls require us to
specify details about the filesystem, which we'd rather auto-detect.

Finally, ensure to cleanly detach volumes if the mounting fails.