Support VolumeDriver.Create options
This adds support for volume driver options and, in the EBS driver,
recognizes --volume_id=x, for situations in which the EBS volume ID
is different than the Docker volume name.  This is common in Docker
Compose files, particularly in version 2 of the file format; e.g.:

    myservice:
        volumes:
            myvolume:/var/blah/blah
    myvolume:
        driver: blocker
        driver_opts:
            volume_id: vol-913a5718

In this case, we refer to the volume as 'myvolume' throughout the
Compose file, however its actual EBS volume ID is vol-913a5718.

The driver still uses the name as the default if this option isn't
provided.